+++ test/audits/image-alt-text/opportunity-handler.test.js
@@ -91,7 +91,7 @@ describe('Image Alt Text Opportunity Handler', () => {
           { pageUrl: '/page2', src: 'image2.jpg' },
           { pageUrl: '/page3', src: 'image1.svg', blob: 'blob' },
         ],
-        presentationalImagesCount: 0,
+        decorativeImagesCount: 0,
       },
     };
 
@@ -138,7 +138,7 @@ describe('Image Alt Text Opportunity Handler', () => {
       data: sinon.match({
         projectedTrafficLost: sinon.match.number,
         projectedTrafficValue: sinon.match.number,
-        presentationalImagesCount: 0,
+        decorativeImagesCount: 0,
         dataSources: [DATA_SOURCES.RUM, DATA_SOURCES.SITE, DATA_SOURCES.AHREFS, DATA_SOURCES.GSC],
       }),
     }));
@@ -419,7 +419,7 @@ describe('Image Alt Text Opportunity Handler', () => {
     expect(altTextOppty.setData).to.have.been.calledWith({
       projectedTrafficLost: 0,
       projectedTrafficValue: 0,
-      presentationalImagesCount: 0,
+      decorativeImagesCount: 0,
       dataSources: [DATA_SOURCES.RUM, DATA_SOURCES.SITE, DATA_SOURCES.AHREFS, DATA_SOURCES.GSC],
     });