+++ CHANGELOG.md
@@ -1,3 +1,122 @@
+# [1.99.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.98.0...v1.99.0) (2025-06-09)
+
+
+### Features
+
+* adding projected conversion value in forms opportunities ([#940](https://github.com/adobe/spacecat-audit-worker/issues/940)) ([28beef4](https://github.com/adobe/spacecat-audit-worker/commit/28beef40833b327d14c389a7cce0d1aacf262d89))
+
+# [1.98.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.11...v1.98.0) (2025-06-09)
+
+
+### Features
+
+* form source will be sent as camel case from mystique ([#939](https://github.com/adobe/spacecat-audit-worker/issues/939)) ([98ade4a](https://github.com/adobe/spacecat-audit-worker/commit/98ade4af1c75b87dcc901892464e708c50f5f79e))
+
+## [1.97.11](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.10...v1.97.11) (2025-06-08)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#938](https://github.com/adobe/spacecat-audit-worker/issues/938)) ([342c353](https://github.com/adobe/spacecat-audit-worker/commit/342c3534ef899bbbcf34475aece8073229d3261e))
+
+## [1.97.10](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.9...v1.97.10) (2025-06-07)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#937](https://github.com/adobe/spacecat-audit-worker/issues/937)) ([e72b304](https://github.com/adobe/spacecat-audit-worker/commit/e72b304b1875435cec791756e4e1ed7a70ac0289))
+
+## [1.97.9](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.8...v1.97.9) (2025-06-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-http-utils to v1.14.1 ([#935](https://github.com/adobe/spacecat-audit-worker/issues/935)) ([79f5e08](https://github.com/adobe/spacecat-audit-worker/commit/79f5e086ca35b3ed06e9a95775c8e0689f354b78))
+
+## [1.97.8](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.7...v1.97.8) (2025-06-04)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#930](https://github.com/adobe/spacecat-audit-worker/issues/930)) ([dc2db0e](https://github.com/adobe/spacecat-audit-worker/commit/dc2db0eeeb91baf9003f5e60574361d8716e0d67))
+
+## [1.97.7](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.6...v1.97.7) (2025-06-03)
+
+
+### Bug Fixes
+
+* **forms:** use form source to identify oppty to update guidance ([#917](https://github.com/adobe/spacecat-audit-worker/issues/917)) ([1c3da05](https://github.com/adobe/spacecat-audit-worker/commit/1c3da053472f4bae8f4a3be3206e65aeffe7f0d3))
+
+## [1.97.6](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.5...v1.97.6) (2025-06-03)
+
+
+### Bug Fixes
+
+* **forms-a11y:** update accessibility data processing and opportunity creation ([#924](https://github.com/adobe/spacecat-audit-worker/issues/924)) ([6199178](https://github.com/adobe/spacecat-audit-worker/commit/61991784b484a86fa74b426e36f4b04da0bbe016))
+
+## [1.97.5](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.4...v1.97.5) (2025-06-03)
+
+
+### Bug Fixes
+
+* detect language from url ([#925](https://github.com/adobe/spacecat-audit-worker/issues/925)) ([3833206](https://github.com/adobe/spacecat-audit-worker/commit/383320655919ed000a7d7df533a5591de512781d))
+
+## [1.97.4](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.3...v1.97.4) (2025-06-03)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-ahrefs-client to v1.8.1 ([#926](https://github.com/adobe/spacecat-audit-worker/issues/926)) ([194f936](https://github.com/adobe/spacecat-audit-worker/commit/194f93616ef587f8f8e49119d2990324b8f24f39))
+
+## [1.97.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.2...v1.97.3) (2025-06-02)
+
+
+### Bug Fixes
+
+* preflight response bottlenecks ([#919](https://github.com/adobe/spacecat-audit-worker/issues/919)) ([3acfaed](https://github.com/adobe/spacecat-audit-worker/commit/3acfaedcf0b5dba55cc25da28577613df4486b16))
+
+## [1.97.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.1...v1.97.2) (2025-06-02)
+
+
+### Bug Fixes
+
+* handling edge cases for calculating rates of forms ([#914](https://github.com/adobe/spacecat-audit-worker/issues/914)) ([db68a47](https://github.com/adobe/spacecat-audit-worker/commit/db68a473511851a7d08f4b349850cc903a34646a))
+
+## [1.97.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.97.0...v1.97.1) (2025-06-02)
+
+
+### Bug Fixes
+
+* preflight internal links identify ([#915](https://github.com/adobe/spacecat-audit-worker/issues/915)) ([76b8f4b](https://github.com/adobe/spacecat-audit-worker/commit/76b8f4bd4f10a42e99ffecbfd5ba804e114ab139))
+
+# [1.97.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.96.4...v1.97.0) (2025-06-02)
+
+
+### Features
+
+* Filter out decorative (almost all) images from audit suggestions ([#923](https://github.com/adobe/spacecat-audit-worker/issues/923)) ([478a053](https://github.com/adobe/spacecat-audit-worker/commit/478a0539662ac1f9da34af1fc8cf74f288b5c766))
+
+## [1.96.4](https://github.com/adobe/spacecat-audit-worker/compare/v1.96.3...v1.96.4) (2025-05-31)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#921](https://github.com/adobe/spacecat-audit-worker/issues/921)) ([0a4ba2b](https://github.com/adobe/spacecat-audit-worker/commit/0a4ba2b7d6ba42e5a6d03f550842b476bbe54d14))
+
+## [1.96.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.96.2...v1.96.3) (2025-05-31)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#920](https://github.com/adobe/spacecat-audit-worker/issues/920)) ([793cd3c](https://github.com/adobe/spacecat-audit-worker/commit/793cd3ca9575ab174bd5d7b56c4af17032f4a9cc))
+
+## [1.96.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.96.1...v1.96.2) (2025-05-30)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-rum-api-client to v2.28.0 ([#918](https://github.com/adobe/spacecat-audit-worker/issues/918)) ([55a5e0c](https://github.com/adobe/spacecat-audit-worker/commit/55a5e0cc0a1903e98f6c06b97bc31a09c4fdf163))
+
 ## [1.96.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.96.0...v1.96.1) (2025-05-29)